requirements.txt
# Core
python-dotenv>=1.0.0
pyyaml>=6.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
jsonlines>=4.0.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=8.0.0

# Embeddings & ML (Stage 1 & 2)
sentence-transformers>=3.0.0
torch>=2.0.0
faiss-cpu>=1.7.0

# Database
pymysql>=1.1.0
sqlalchemy>=2.0.0

# LLM (Stage 3)
anthropic>=0.40.0

# Visualization & Validation
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
jsonschema>=4.17.0

# Testing
pytest>=7.0.0
